docs: Update patch instructions
The XXD patches were a nice simple solution, but ultimately swapping the textures ended up not being a very good solution by themselves; ideally I would've also updated the meshes, but this gets quite complex. First, the meshes are vertex meshes, so updating the mesh doesn't update the animations; they'd all need to be updated by hand, frame-by-frame. Then, there isn't a straightforward way to update the meshes in the UPX file. Some of the Unreal package files can be updated with the Deus Ex SDK, but not DeusExCharacters.u because the sources aren't included in the SDK (according to the SDK documentation). Editing the package file manually isn't feasible because the mesh size would change, meaning all of the offsets in the package file would need to be updated. And after all that, who knows if/how well the mesh would work.

Instead, a small script was created to 'mask' textures. It's far from ideal but in the end it's a pretty straightforward solution and relatively easy to apply the patch, only requiring Node.js to be installed. If there was any interest, the script could even be packaged into a binary using nexe, but it's probably overkill for a very specific patch like this.